Rename and move python tests for a proper dir
