+++ tests/test_chemicalproperties.py
