+++ tests/test_equilibriumproblem.py
