+++ tests/test_equilibriumsolver.py
